add licence header
