+++ client/src/main/java/cn/vika/client/api/datasheet/AttachmentApi.java
@@ -1,3 +1,21 @@
+/*
+ * Copyright (C) 2021 vikadata
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program; if not, write to the Free Software Foundation, Inc.,
+ * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+ */
+
 package cn.vika.client.api.datasheet;
 
 import java.io.ByteArrayOutputStream;
@@ -7,10 +25,10 @@ import java.util.Map;
 import java.util.UUID;
 
 import cn.vika.client.api.exception.ApiException;
-import cn.vika.client.api.http.AbstractApi;
 import cn.vika.client.api.http.ApiHttpClient;
 import cn.vika.client.api.model.AbstractModel;
 import cn.vika.client.api.model.HttpResult;
+import cn.vika.client.api.http.AbstractApi;
 import cn.vika.core.http.GenericTypeReference;
 import cn.vika.core.http.HttpHeader;
 import cn.vika.core.http.HttpMediaType;
