+++ client/src/main/java/cn/vika/client/api/models/RecordPageInfo.java
@@ -1,4 +1,22 @@
-package cn.vika.client.datasheet.model;
+/*
+ * Copyright (C) 2021 vikadata
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program; if not, write to the Free Software Foundation, Inc.,
+ * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+ */
+
+package cn.vika.client.api.models;
 
 import java.util.HashMap;
 import java.util.List;
@@ -56,11 +74,6 @@ public class RecordPageInfo extends AbstractModel {
         this.records = records;
     }
 
-    /**
-     * request query param to map
-     *
-     * @param map
-     */
     @Override
     public void toMap(HashMap<String, String> map, String prefix) {
