+++ client/src/main/java/cn/vika/client/api/models/RecordResponse.java
@@ -16,26 +16,30 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-package cn.vika.client.api.datasheet;
+package cn.vika.client.api.models;
+
+import java.util.HashMap;
 
 import cn.vika.client.api.model.AbstractModel;
-import cn.vika.client.api.model.HttpResult;
-import cn.vika.core.http.GenericTypeReference;
 
 /**
- * attachment api
+ * create record response
  *
  * @author Zoe Zheng
- * @date 2020-12-17 16:15:54
+ * @date 2020-12-17 11:22:01
  */
-public interface IAttachmentApi {
+public class RecordResponse extends AbstractModel {
 
-    /**
-     * upload datasheet attachment
-     *
-     * @param params add attachment data
-     * @param responseType response type
-     * @return responseType
-     */
-    <T> T uploadAttachment(String datasheetId, AbstractModel params, GenericTypeReference<HttpResult<T>> responseType) throws Exception;
+    private RecordDetail[] records;
+
+    public RecordDetail[] getRecords() {
+        return records;
+    }
+
+    public void setRecords(RecordDetail[] records) {
+        this.records = records;
+    }
+
+    @Override
+    public void toMap(HashMap<String, String> map, String prefix) {}
 }
