--- client/src/main/java/cn/vika/client/datasheet/model/FieldKey.java
@@ -1,18 +1,0 @@
-package cn.vika.client.datasheet.model;
-
-/**
- * FieldKey enum
- *
- * @author Zoe Zheng
- * @date 2020-12-16 14:10:31
- */
-public enum FieldKey {
-    /**
-     * find fields with name
-     */
-    Name,
-    /**
-     * find fields with id
-     */
-    ID
-}
