--- client/src/main/java/cn/vika/client/datasheet/model/Order.java
@@ -1,18 +1,0 @@
-package cn.vika.client.datasheet.model;
-
-/**
- * sort order enum
- *
- * @author Zoe Zheng
- * @date 2020-12-16 14:10:31
- */
-public enum Order {
-    /**
-     * sort records descending
-     */
-    DESC,
-    /**
-     * sort records ascending
-     */
-    ASC
-}
