--- client/src/main/java/cn/vika/client/datasheet/model/RecordResponse.java
@@ -1,32 +1,0 @@
-package cn.vika.client.datasheet.model;
-
-import java.util.HashMap;
-
-import cn.vika.client.api.model.AbstractModel;
-
-/**
- * create record response
- *
- * @author Zoe Zheng
- * @date 2020-12-17 11:22:01
- */
-public class RecordResponse extends AbstractModel {
-
-    private RecordDetail[] records;
-
-    public RecordDetail[] getRecords() {
-        return records;
-    }
-
-    public void setRecords(RecordDetail[] records) {
-        this.records = records;
-    }
-
-    /**
-     * request query param to map
-     *
-     * @param map
-     */
-    @Override
-    public void toMap(HashMap<String, String> map, String prefix) {}
-}
